Split Network.Socket module, create Network.Socket.IO
The non-streaming array IO APIs go into the IO module. More IO based
non-streaming APIs can later be exported from this module.
