+++ src/Streamly/Internal/Network/Socket.hs
@@ -17,7 +17,7 @@ module Streamly.Internal.Network.Socket
     (
     SockSpec (..)
     -- * Use a socket
-    , handleWithM
+    , SKIO.handleWithM
     , handleWith
 
     -- * Accept connections
@@ -56,7 +56,6 @@ module Streamly.Internal.Network.Socket
     , fromBytes
 
     -- -- * Array Write
-    , writeChunk
     , writeChunks
     , writeChunksWithBufferOf
     , writeStrings
@@ -65,25 +64,14 @@ module Streamly.Internal.Network.Socket
     )
 where
 
-import Control.Concurrent (threadWaitWrite, rtsSupportsBoundThreads)
-import Control.Monad.Catch (MonadCatch, finally, MonadMask)
+import Control.Monad.Catch (MonadCatch)
 import Control.Monad.IO.Class (MonadIO(..))
-import Control.Monad (when)
 import Data.Word (Word8)
-import Foreign.ForeignPtr (withForeignPtr)
-import Foreign.ForeignPtr.Unsafe (unsafeForeignPtrToPtr)
-import Foreign.Ptr (minusPtr, plusPtr, Ptr, castPtr)
 import Foreign.Storable (Storable(..))
-import GHC.ForeignPtr (mallocPlainForeignPtrBytes)
 import Network.Socket
        (Socket, SocketOption(..), Family(..), SockAddr(..),
         ProtocolNumber, withSocketsDo, SocketType(..), socket, bind,
-        setSocketOption, sendBuf, recvBuf)
-#if MIN_VERSION_network(3,1,0)
-import Network.Socket (withFdSocket)
-#else
-import Network.Socket (fdSocket)
-#endif
+        setSocketOption)
 import Prelude hiding (read)
 
 import qualified Network.Socket as Net
@@ -103,19 +91,9 @@ import qualified Streamly.Internal.Memory.Array as IA
 import qualified Streamly.Memory.Array as A
 import qualified Streamly.Internal.Memory.ArrayStream as AS
 import qualified Streamly.Internal.Memory.Array.Types as A
+import qualified Streamly.Internal.Network.Socket.IO as SKIO
 import qualified Streamly.Prelude as S
 import qualified Streamly.Internal.Data.Stream.StreamD.Type as D
-
--- | @'handleWithM' socket act@ runs the monadic computation @act@ passing the
--- socket handle to it.  The handle will be closed on exit from 'handleWithM',
--- whether by normal termination or by raising an exception.  If closing the
--- handle raises an exception, then this exception will be raised by
--- 'handleWithM' rather than any exception raised by 'act'.
---
--- @since 0.7.0
-{-# INLINE handleWithM #-}
-handleWithM :: (MonadMask m, MonadIO m) => (Socket -> m ()) -> Socket -> m ()
-handleWithM f sk = finally (f sk) (liftIO (Net.close sk))
 
 -- | Like 'handleWithM' but runs a streaming computation instead of a monadic
 -- computation.
@@ -205,82 +183,6 @@ connections :: MonadAsync m => Int -> SockSpec -> SockAddr -> SerialT m Socket
 connections tcpListenQ spec addr = fst <$> recvConnectionTuplesWith tcpListenQ spec addr
 
 -------------------------------------------------------------------------------
--- Array IO (Input)
--------------------------------------------------------------------------------
-
-{-# INLINABLE readArrayUptoWith #-}
-readArrayUptoWith
-    :: (h -> Ptr Word8 -> Int -> IO Int)
-    -> Int
-    -> h
-    -> IO (Array Word8)
-readArrayUptoWith f size h = do
-    ptr <- mallocPlainForeignPtrBytes size
-    -- ptr <- mallocPlainForeignPtrAlignedBytes size (alignment (undefined :: Word8))
-    withForeignPtr ptr $ \p -> do
-        n <- f h p size
-        let v = Array
-                { aStart = ptr
-                , aEnd   = p `plusPtr` n
-                , aBound = p `plusPtr` size
-                }
-        -- XXX shrink only if the diff is significant
-        -- A.shrinkToFit v
-        return v
-
--- | Read a 'ByteArray' from a file handle. If no data is available on the
--- handle it blocks until some data becomes available. If data is available
--- then it immediately returns that data without blocking. It reads a maximum
--- of up to the size requested.
-{-# INLINABLE readArrayOf #-}
-readArrayOf :: Int -> Socket -> IO (Array Word8)
-readArrayOf = readArrayUptoWith recvBuf
-
--------------------------------------------------------------------------------
--- Array IO (output)
--------------------------------------------------------------------------------
-
-waitWhen0 :: Int -> Socket -> IO ()
-waitWhen0 0 s = when rtsSupportsBoundThreads $
-#if MIN_VERSION_network(3,1,0)
-    withFdSocket s $ \fd -> threadWaitWrite $ fromIntegral fd
-#elif MIN_VERSION_network(3,0,0)
-    fdSocket s >>= threadWaitWrite . fromIntegral
-#else
-    let fd = fdSocket s in threadWaitWrite $ fromIntegral fd
-#endif
-waitWhen0 _ _ = return ()
-
-sendAll :: Socket -> Ptr Word8 -> Int -> IO ()
-sendAll _ _ len | len <= 0 = return ()
-sendAll s p len = do
-    sent <- sendBuf s p len
-    waitWhen0 sent s
-    -- assert (sent <= len)
-    when (sent >= 0) $ sendAll s (p `plusPtr` sent) (len - sent)
-
-{-# INLINABLE writeArrayWith #-}
-writeArrayWith :: Storable a
-    => (h -> Ptr Word8 -> Int -> IO ())
-    -> h
-    -> Array a
-    -> IO ()
-writeArrayWith _ _ arr | A.length arr == 0 = return ()
-writeArrayWith f h Array{..} = withForeignPtr aStart $ \p ->
-    f h (castPtr p) aLen
-    where
-    aLen =
-        let p = unsafeForeignPtrToPtr aStart
-        in aEnd `minusPtr` p
-
--- | Write an Array to a file handle.
---
--- @since 0.7.0
-{-# INLINABLE writeChunk #-}
-writeChunk :: Storable a => Socket -> Array a -> IO ()
-writeChunk = writeArrayWith sendAll
-
--------------------------------------------------------------------------------
 -- Stream of Arrays IO
 -------------------------------------------------------------------------------
 
@@ -309,7 +211,7 @@ toChunksWithBufferOf size h = D.fromStreamD (D.Stream step ())
     where
     {-# INLINE_LATE step #-}
     step _ _ = do
-        arr <- liftIO $ readArrayOf size h
+        arr <- liftIO $ SKIO.readArrayOf size h
         return $
             case A.length arr of
                 0 -> D.Stop
@@ -337,7 +239,7 @@ readChunksWithBufferOf = Unfold step return
     where
     {-# INLINE_LATE step #-}
     step (size, h) = do
-        arr <- liftIO $ readArrayOf size h
+        arr <- liftIO $ SKIO.readArrayOf size h
         return $
             case A.length arr of
                 0 -> D.Stop
@@ -411,7 +313,7 @@ read = UF.supplyFirst readWithBufferOf A.defaultChunkSize
 {-# INLINE fromChunks #-}
 fromChunks :: (MonadIO m, Storable a)
     => Socket -> SerialT m (Array a) -> m ()
-fromChunks h = S.mapM_ (liftIO . writeChunk h)
+fromChunks h = S.mapM_ (liftIO . SKIO.writeChunk h)
 
 -- | Write a stream of arrays to a socket.  Each array in the stream is written
 -- to the socket as a separate IO request.
@@ -419,7 +321,7 @@ fromChunks h = S.mapM_ (liftIO . writeChunk h)
 -- @since 0.7.0
 {-# INLINE writeChunks #-}
 writeChunks :: (MonadIO m, Storable a) => Socket -> Fold m (Array a) ()
-writeChunks h = FL.drainBy (liftIO . writeChunk h)
+writeChunks h = FL.drainBy (liftIO . SKIO.writeChunk h)
 
 -- | @writeChunksWithBufferOf bufsize socket@ writes a stream of arrays
 -- to @socket@ after coalescing the adjacent arrays in chunks of @bufsize@.
